users/serializers.py
from rest_framework import serializers
from django.contrib.auth import get_user_model

CustomUser = get_user_model()

class CustomUserSerializer(serializers.ModelSerializer):
    class Meta:
        model = CustomUser
        fields = '__all__'


    def create(self, validated_data):
        user = CustomUser.objects.create_user(**validated_data)
        return user